Fixes comms timeout at start of match
The stream read function was returning the time of the message as a
text string rather than as an integer.  This caused delay calculations
to be wrong, and the whole thing just timed out.  By making it return
an int, the calculations can now be performed correctly.

Also removed the parameter from the match.start function.  I think
this is legacy code that should have been removed before

+++ web/stream.js
@@ -60,7 +60,7 @@ Comms = function () {
 							if(m) {
 								var c = m.get('count');
 								if(++counter != c) messageBoard.appendText(' [C:'+counter+':'+c+']'); //Should catch out of sequence counts
-								messageCallback(m.get('time'),m.get('text'));
+								messageCallback(m.get('time').toInt(),m.get('text'));
 							}
 						}
 						reader.post({oid:oid}); //Queue up next request
